- Added Facet::LoadXML
- Added Geometry::InsertXML
- File.cpp methods cleanup
- Structure name reading to new structures
- Few int->size_t and char->const char changes
- PARFileList made vector of string from a class
- Worker: unified load and insert

+++ ProfilePlotter.cpp
@@ -469,7 +469,7 @@ void ProfilePlotter::SetViews(std::vector<int> views,int mode) {
 std::vector<int> ProfilePlotter::GetViews() {
 	std::vector<int>v;
 	v.reserve(nbView);
-	for (size_t i = 0; i < nbView; i++)
+	for (int i = 0; i < nbView; i++)
 		v.push_back(views[i]->userData1);
 	return v;
 }
